requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==5.5.3
openai==1.3.4
qdrant-client==1.8.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
cryptography>=41.0.0
tiktoken==0.5.2
pyyaml==6.0.1
slowapi==0.1.9
markdown==3.7
beautifulsoup4==4.12.3